tripletex/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
requests>=2.32.0
google-genai>=1.0.0
openai>=1.0.0
pypdf>=4.0.0
google-cloud-storage>=2.18.0
